+++ src/theme/theme.types.ts
@@ -358,6 +358,7 @@ export type ThemeColors = {
   flagBorder: string;
   /** gray2 */
   fundExchange: string;
+
   /** complementaryGreen1 */
   generationSavingsTimelineColor1: string;
   /** complementaryPink1 */
@@ -368,6 +369,8 @@ export type ThemeColors = {
   generationSavingsTimelineColor4: string;
   /** gray100 */
   keyFiguresBackground: string;
+  /** grey400 */
+  graphCrosshair: string;
   /** gray5 */
   graphVolume: string;
   /** gray4 */
@@ -660,6 +663,12 @@ export type ThemeColors = {
   sfdrArticle9: string;
   /** pink100 */
   popularBadgeBackground: string;
+  /** gray800 */
+  periodScrollButtonBackground: string;
+  /** gray100 */
+  periodScrollButtonIcon: string;
+  /** gray200 */
+  technicalAnalysisPlotBands: string;
   /** Join Shareville Illustration */
   joinSharevilleIllustration: string;
   /** Join Shareville Badge */
@@ -668,6 +677,8 @@ export type ThemeColors = {
   joinSharevilleBanner: string;
   /** Monthly Savings Transfer Type Inactive */
   monthlySavingsTransferTypeInactive: string;
+  /** day select input inactive */
+  daySelectInputTextInactive: string;
   /** @deprecated * gray3 */
   disabled: string;
   /** @deprecated * complementaryBlue2 */